(feat): implement root layout with header component and error boundary
Integrated the root layout with a global Header component displaying a logout button for authenticated users. Added an ErrorBoundary at the root level to catch and handle component errors gracefully. Updated main.tsx to use RouterProvider with the TanStack router instance, completing the router integration. Removed App.tsx as it's no longer needed with router-based architecture. Enhanced environment validation with proper error handling for production. The Header is responsive with dark mode support. Kept VITE_ENABLE_LEARN_PROMPT feature flag for controlling the Learn 2 Prompt button.
