+++ webapp/frontend/src/components/ErrorBoundary.tsx
@@ -1,0 +1,56 @@
+import React, { Component, ReactNode, ErrorInfo } from 'react';
+
+interface Props {
+  children: ReactNode;
+  fallback?: ReactNode;
+}
+
+interface State {
+  hasError: boolean;
+  error: Error | null;
+}
+
+export class ErrorBoundary extends Component<Props, State> {
+  constructor(props: Props) {
+    super(props);
+    this.state = { hasError: false, error: null };
+  }
+
+  static getDerivedStateFromError(error: Error): State {
+    return { hasError: true, error };
+  }
+
+  componentDidCatch(error: Error, errorInfo: ErrorInfo): void {
+    console.error('ErrorBoundary caught:', error, errorInfo);
+  }
+
+  handleReset = (): void => {
+    this.setState({ hasError: false, error: null });
+  };
+
+  render(): ReactNode {
+    if (this.state.hasError) {
+      return (
+        this.props.fallback || (
+          <div className="flex items-center justify-center min-h-screen bg-white dark:bg-gray-900">
+            <div className="text-center px-4">
+              <h1 className="text-4xl font-bold text-gray-900 dark:text-white mb-4">Oops!</h1>
+              <p className="text-lg text-gray-600 dark:text-gray-400 mb-6">
+                Something went wrong. Our team has been notified.
+              </p>
+              <button
+                onClick={this.handleReset}
+                className="px-6 py-3 bg-blue-600 hover:bg-blue-700 text-white rounded font-medium transition-colors"
+                aria-label="Retry"
+              >
+                Try Again
+              </button>
+            </div>
+          </div>
+        )
+      );
+    }
+
+    return this.props.children;
+  }
+}
