+++ webapp/frontend/src/main.tsx
@@ -1,9 +1,10 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { QueryClientProvider } from '@tanstack/react-query';
+import { RouterProvider } from '@tanstack/react-router';
 import { queryClient } from './lib/queryClient';
 import { AuthProvider } from './contexts/AuthContext';
-import App from './App';
+import { router } from './routes';
 
 const rootElement = document.getElementById('root');
 if (!rootElement) throw new Error('Root element not found!');
@@ -12,7 +13,7 @@ createRoot(rootElement).render(
   <StrictMode>
     <QueryClientProvider client={queryClient}>
       <AuthProvider>
-        <App />
+        <RouterProvider router={router} />
       </AuthProvider>
     </QueryClientProvider>
   </StrictMode>
