+++ webapp/frontend/src/pages/LoginPage.tsx
@@ -1,4 +1,4 @@
-import { LoginModal } from '../components/LoginModal';
+import LoginModal from '../components/LoginModal';
 
 /**
  * Login page component
